280619_0313PM_addingCn1Changes

+++ CodenameOne/src/com/codename1/ui/util/UITimer.java
@@ -38,7 +38,7 @@ public class UITimer {
     private Runnable internalRunnable;
     private Form bound;
     private long lastEllapse;
-    private int ms;
+    private long ms;
     private boolean repeat;
     private Internal i = new Internal();
